Added network_measures, location_crosstab and location_information to inspecot as methods
